// src/presentation/texts-reservoir/core/orders/interface.ts
export interface ICoreOrdersTextConstants {
  WELCOME_MODAL: {
    HEADING: string;
    SUB_HEADING: string;
    BUTTON_TEXT: string;
  };
  TAB_NAMES: {
    ALL: string;
    DRAFTS: string;
  };
  CTA_BUTTON: {
    TEXT: string;
  };
  FILTER_NAMES: {
    CHANNEL_NAME: string;
    ORDER_STATUS: string;
    DELIVERY_STATUS: string;
    INVOICING_STATUS: string;
    PAYMENT_STATUS: string;
    CHARGEBACK_STATUS: string;
  };
  VIEW: {
    ORDER_DETAILS: {
      TITLE: string;
      ATTACHMENT_TITLE: string;
      ATTACHMENTS_INFO: string;
      UPLOAD_BUTTON_TEXT: string;
      LIST: {
        ORDER_NAME: string;
        CHANNEL: string;
        PO_DATE: string;
        REQUESTED_DATE: string;
        SHIP_TO_ADDRESS: string;
        NET_DAYS: string;
        REQUESTED_DELIVERY: string;
        EARLIEST_SHIP: string;
        LATEST_SHIP: string;
        REQUESTED_SHIP: string;
        LATEST_DELIVERY: string;
        SHIPMENT_TRACKING: string;
        NET_DUE_DAYS: string;
        DISCOUNT_DUE_DAYS: string;
        DISCOUNT_PERCENTAGE: string;
        TERMS_TYPE: string;
        TERMS_START_DATE: string;

        DATA_ERROR: string;
      };
    };
    COMMENTS: {
      HEADING: string;
      ADD_INFO: string;
      ADD_BUTTON_TEXT: string;
      MENTIONS_PLACEHOLDER: string;
      DELETE_COMMENT_BUTTON_TEXT: string;
    };
  };
}